backend/pyproject.toml
[tool.poetry]
name = "skylock"
version = "0.1.0"
description = ""
authors = ["akempins <01178535@pw.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
pylint = "^3.3.1"
fastapi = "^0.115.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
black = "^24.10.0"
alembic = "^1.13.3"
pytest-cov = "^5.0.0"
argon2-cffi = "^23.1.0"
pathlib = "^1.0.1"
python-multipart = "^0.0.17"
gunicorn = "^23.0.0"
jinja2 = "^3.1.4"
pyotp = "^2.9.0"
redis = "^5.2.1"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.166.0"
mypy = "^1.15.0"
dramatiq = {extras = ["watch"], version = "^1.14"}
slowapi = "^0.1.9"
vulture = "^2.14"
types-python-jose = "^3.4.0.20250516"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:datetime.datetime.utcnow\\(\\) is deprecated and scheduled for removal in a future version:DeprecationWarning",
    "ignore:Please use `import python_multipart` instead.:PendingDeprecationWarning",
]
addopts = "--cov --cov-report=term-missing"

[tool.coverage.run]
source = ["skylock"]

[tool.pylint]
disable = [
    "C0114", # (missing-module-docstring)
    "C0115", # (missing-class-docstring)
    "C0116", # (missing-function-docstring)
    "R0903", # (too-few-public-methods)
    "C0301", # (line-to-long)
    "E1136", # (object-not-subscriptable)
    "W0511", # (TODO)
    "R0913", # (too-many-arguments)
    "R0917", # (too-many-positional-arguments)
]

[tool.black]
line-length = 100
